pyproject.toml
[tool.poetry]
name = "ims"
version = "0.1.0"
description = ""
authors = ["Jaefer-Kemal <jaeferkemal@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.2"
djangorestframework = "^3.15.2"
pillow = "^10.4.0"
psycopg2 = "^2.9.9"
django-filter = "^24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
